experiments/requirements.txt
numpy
Pillow
colour-science
pypng